+++ tests/run_tests.py
@@ -45,6 +45,7 @@ from mcfm_test import *
 from hessian_finite_differences import * 
 from eam_calculator_forces_and_hessian import *
 from pair_potential_calculator import *
+from polydisperse_calculator import *
 
 if sys.version_info.major > 2:
     from test_c2d import *